requirements.txt
hyperliquid-python-sdk==0.9.3
eth-account>=0.8.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
python-telegram-bot>=20.0
aiohttp>=3.9.0
requests>=2.31.0
